feat: Implement exponential backoff to GeminiLLM, and enable it by default
Exponential backoff is enabled by default. It start with 5 seconds and ^2 backoff for
subsequent retries with max 60s delay. As the result, it fires at 5s, 10s, 20s, 40s, 60s.

Usage: It is enabled by default, but is configurable during agent declaration:

```python
root_agent = Agent(
  model=Gemini(
    model='gemini-2.0-flash',
    retry_config=RetryConfig(initial_delay_sec=60, max_retries=3)
  ),
  ...
)
```

Note: This config cannot be added to RunConfig. Although there are similar configurations,
RunConfig is only available in invocation_context, which is not available to BaseLLM and
any derived LLM classes.

Tested locally:

```bash
 The description about you is "Checks if input is valid using predefined tools"'
[logging_plugin]    Available Tools: ['check_valid_input', 'check_valid_input2']
2025-07-21 18:00:07.767904
2025-07-21 18:00:12.776910
2025-07-21 18:00:22.792078
2025-07-21 18:00:42.817873
2025-07-21 18:01:22.856147
[logging_plugin] 🧠 LLM ERROR
[logging_plugin]    Agent: check_input
[logging_plugin]    Error: 503 None. {}

```

PiperOrigin-RevId: 784029405